Capstone Project/Anshika Singla/Capstone_Final/requirements.txt
fundus
py2neo
streamlit
transformers
requests
google-generativeai
tensorflow
ntlk
langchain_community
elasticsearch
numpy

